# doc/api.yaml
openapi : 3.0.0
info:
  title: WASAPhoto
  description: |
    This API allows people to keep up with their friend's life by watching and eventually liking their posted photos. Furthermore you have the chance to post your favorite moments and to get love from you most affectionate peolple through likes on your posting.
  version: 0.0.1
...
paths:
  /login/:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  $ref: "#/components/schemas/username"
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: 
                    $ref: "#/components/schemas/nameUser"  
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /users/{userid}:
    parameters:
    -  $ref: "#/components/parameters/username"
    put:
      tags: [User]
      summary: Change Username
      description: |
        Using this command to set the username of a profile if not existing, otherwise,
        the command updates it
      operationId: setMyUserName
      requestBody:
        description: User Id
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
        required: true
      responses:
        "201":
          description: Username set
            
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: 
                    $ref: "#/components/schemas/Name" 
        "400":
          $ref: "#/components/responses/BadRequest"
    
    

      
    delete:
      tags: [User]
      summary: Delete my profile
      description: |
        Using this command if a certain user wants to delete his/her own profile
      operationId: deleteMyProfile
      
      responses:
        "204":
          description: User deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /users/{userid}/banned/{banneduser}:
    parameters:
    - $ref: "#/components/parameters/username"
      
    - $ref: "#/components/parameters/banned"
      
    put:
      tags: [User]
      summary: Ban another user
      description: A certain user can always ban another user in order to not by visible by him/her
      operationId: banUser
      requestBody:
        description: Id of the banned user
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
                
      responses:
        "201":
          description: The user has been correctly banned
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    ArrayofPhotos:
                      type: array
                      minItems: 0
                      maxItems: 1000 
                      items:
                        $ref: "#/components/schemas/Name"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    
    delete:
      tags: [User]
      summary: Unban user
      description: |
        Delete a user from the list of the banned people
      operationId: unbanUser
      responses:
        "204":
          description: User correctly removed from banned list
          
        "404": 
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /users/{userid}/followed/{followeduser}:
    parameters:
    - $ref: "#/components/parameters/username"
      
    - $ref: "#/components/parameters/followed"
      

    
    put:
      tags: [Follow]
      summary: Add a user to the list of followed people
      description: Using this command we can add this 
        profile to the list of profiles a certain user is following
      operationId: followUser
      requestBody:
        description: Id of followed user
        content:
          application/json:
            schema:
              type: object
              properties:
                Name:
                  $ref: "#/components/schemas/Name"
      responses:
        "201":
          description: User correctly added to the list of followed users
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    ArrayofPhotos:
                      type: array
                      minItems: 0
                      maxItems: 1000 
                      items:
                        $ref: "#/components/schemas/Name"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

    
    delete:
      tags: [Follow]
      summary: Stop following a user
      description: Using this command a user can stop following another user
      operationId: unfollowUser
      responses:
        "204":
          description: User correctly unfollowed
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /users/{userid}/photos/:
    parameters:
      - $ref: "#/components/parameters/username"
    
    post:
      tags: [Photo]
      summary: Add a new photo to a certain profile
      description: |
        Using this command if a certain user wants to delete his/her own profile
      operationId: uploadPhoto
      requestBody:
        description: Id of the added photo
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/PhotoId"
      responses:
        "201":
          description: Photo added
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  parameters:
    username: 
      name: username
      in: path
      required: true
      schema:
        type: string
        pattern: 'ˆ.*?$'
        minLength: 3
        maxLength: 16 
    banned:
      name: banned
      in: path
      description: The username of the user that has been banned
      required: true
      schema:
        $ref: "#/components/schemas/Name"
...
    followed:
      name: followed
      in: path
      description: Username of the follower taken int consideration
      required: true
      schema:
        $ref: "#/components/schemas/Name"
...
  schemas:
    Name:
      type: string
      pattern: 'ˆ.*?$'
      minLength: 3
      maxLength: 16
      description: User identification on that social
    PhotoId:
      description: value identifying of a photo
      type: number
      format: int64
...
    Error:
      type: object
      description: general error
      properties:
        code:
          type: string
          description: The code of the error
          pattern: "ˆ.*?$"
          
        message:
          type: string
          description: The message of the error explaining what is happening
          pattern: "ˆ.*?$"
          
      required:
        - code
        - message
...
  securitySchemes:
    bearerAuth:  
      description: bearer authentication          
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid 
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The request conflicts with the current state of the server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: The request cannot be processed because the value already exists in the database
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Operation not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []